PORTFO/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
sqlalchemy-utils
